% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{form_correlation}
\alias{form_correlation}
\title{Form correlation}
\usage{
form_correlation(path = "/raid6/Kevin/supernodeWGS_results/blocks/",
  summary_func = function(x) {     mean(as.numeric(unlist(x))) }, K = 200,
  max_cluster = NA, cores = 1, verbose = T)
}
\arguments{
\item{path}{filepath prefix}

\item{summary_func}{function to summarize each edge}

\item{K}{number of nodes or vector of indicies}

\item{max_cluster}{maximum number of clusters}

\item{cores}{number of cores for parallelization}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Form correlation
}
