% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{form_graph}
\alias{form_graph}
\title{Form graph}
\usage{
form_graph(path = "/raid6/Kevin/supernodeWGS_results/blocks/",
  edge_func = function(x) {     abs(mean(as.numeric(unlist(x)))) > 0.1 },
  K = 200, max_cluster = NA, cores = 1, verbose = T)
}
\arguments{
\item{path}{filepath prefix}

\item{edge_func}{function to determine if there's an edge}

\item{K}{number of nodes or vector of indicies}

\item{max_cluster}{maximum number of clusters}

\item{cores}{number of cores for parallelization}

\item{verbose}{boolean}
}
\value{
\code{igraph} object
}
\description{
Form graph
}
