% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_correlation.R
\name{form_graph_from_correlation}
\alias{form_graph_from_correlation}
\title{Create graph for correlation}
\usage{
form_graph_from_correlation(cor_mat, func = function(x) {     x > 0.15 },
  K = 200)
}
\arguments{
\item{cor_mat}{list of correlation values}

\item{func}{function to determine presence of edge}

\item{K}{number of nodes or vector of indicies}
}
\value{
graph
}
\description{
The input, \code{cor_mat}, follows the format given by \code{form_correlation}.
This is important if \code{K} is a vector of indices, as opposed a positive
integer for how many nodes there are.
}
