% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{form_testvec}
\alias{form_testvec}
\title{Form test statistics}
\usage{
form_testvec(vec, clustering, flag_vec = rep(FALSE, length(vec)),
  path = "/raid6/Kevin/supernodeWGS_results/blocks/", K = 200,
  max_cluster = NA, sparse = F, sumabsv = 4, cores = 1, verbose = T)
}
\arguments{
\item{vec}{vector of test statistics}

\item{clustering}{vector of clustering assignments}

\item{flag_vec}{boolean vector of same length as \code{vec}, where indices with
\code{TRUE} will be omitted from calculation}

\item{path}{filepath}

\item{K}{number of nodes or vector of indicies}

\item{max_cluster}{maximum number of clusters}

\item{sparse}{boolean for sparse PCA}

\item{sumabsv}{paramter for sparse PCA}

\item{cores}{number of cores for parallelization}

\item{verbose}{boolean}

\item{...}{extra parameters for PCA::SPC}
}
\value{
vector
}
\description{
Form test statistics
}
