% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{hmrf}
\alias{hmrf}
\title{Hidden markov random field}
\usage{
hmrf(z, adj, seedindex = rep(0, length(z)), null_mean = 0,
  null_sigma = NA, pthres = 0.05, iter = 100, verbose = FALSE,
  tol = 0.001)
}
\arguments{
\item{z}{a vector of d z-values}

\item{adj}{a dxd adjancey matrix (0,1 entries)}

\item{seedindex}{a (0,1) vector of length d, where 1 means the gene is in group 1.}

\item{null_mean}{the mean of the null distribution}

\item{null_sigma}{the sd of the null distribution}

\item{pthres}{threshold for p-values to serve as an initialization}

\item{iter}{number of iterations}

\item{verbose}{boolean}

\item{tol}{tolerance for values that should be treated as zero in absolute value}
}
\value{
a list
}
\description{
Use a Ising model to update the pvalues according to the graphical structure.
There are two groups, indices in group 0 have mean 0. Indices in group 1 have
mean greater than 0.
}
\details{
In the output, \code{post} refers to the posterior probability of being in
group 1.
}
