% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{index_to_pair}
\alias{index_to_pair}
\title{Index to pair}
\usage{
index_to_pair(val, max_val = 200)
}
\arguments{
\item{val}{numeric}

\item{max_val}{numeric}
}
\value{
numeric
}
\description{
Index to pair
}
