% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Matching two vectors}
\usage{
matching(name1, name2, safety = T)
}
\arguments{
\item{name1}{vector}

\item{name2}{vector}

\item{safety}{boolean}
}
\value{
numeric vector
}
\description{
Returns indicies so \code{all(name1 == name2[matching(name1, name2)])}.
}
