% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{pair_to_index}
\alias{pair_to_index}
\title{Pair to index}
\usage{
pair_to_index(idx1, idx2, max_val = 200)
}
\arguments{
\item{idx1}{numeric}

\item{idx2}{numeric}

\item{max_val}{numeric}
}
\value{
numeric
}
\description{
Pair to index
}
