% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_name.R
\name{screen_name}
\alias{screen_name}
\title{Screen names based on words in certain positions}
\usage{
screen_name(vec, position = 4, stopwords = c("Any", "CodingRegion",
  "FrameshiftRegion", "LoFRegion", "MissenseRegion", "SilentRegion",
  "MissenseHVARDRegionSimple"))
}
\arguments{
\item{vec}{vector of names}

\item{position}{index}

\item{stopwords}{words to remove}
}
\value{
vector of booleans
}
\description{
Screen names based on words in certain positions
}
