% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_name.R
\name{tabulate_names}
\alias{tabulate_names}
\title{Tabulate names based on position}
\usage{
tabulate_names(vec, position = 4)
}
\arguments{
\item{vec}{vector of names}

\item{position}{index}
}
\value{
vector of counts
}
\description{
Tabulate names based on position
}
